requirements.txt
streamlit
cloudinary
Pillow
requests
python-dotenv